% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rg_batch_query}
\alias{rg_batch_query}
\title{Look up vector of points}
\usage{
rg_batch_query(ctx, points, field)
}
\arguments{
\item{ctx}{V8 context created by rg_load_polygons}

\item{points}{list of points}

\item{field}{name of the field that should be returned for the matching polygon}
}
\value{
vector of values of the field chosen in \code{field} for the matching polygon for each point
}
\description{
Look up vector of points
}
