% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rg_load_polygons}
\alias{rg_load_polygons}
\title{Load polygons}
\usage{
rg_load_polygons(data)
}
\arguments{
\item{data}{R list object containing the polygons to be indexed}
}
\value{
V8 context with spatial polygons loaded/indexed
}
\description{
Creates a V8 context and loads the polygons, creating a spatial index
}
