% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rg_query}
\alias{rg_query}
\title{Look up single point}
\usage{
rg_query(ctx, point, field)
}
\arguments{
\item{ctx}{V8 context created by rg_load_polygons}

\item{point}{vector with x/y coordinates for point}

\item{field}{name of the field that should be returned for the matching polygon}
}
\value{
value of the field chosen in \code{field} for the matching polygon
}
\description{
Look up single point
}
